feat: add observability starters
