+++ dubbo-test/dubbo-dependencies-all/pom.xml
@@ -328,7 +328,7 @@
 
         <dependency>
             <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-spring-boot-observability-starter</artifactId>
+            <artifactId>dubbo-spring-boot-observability-otel-zipkin-starter</artifactId>
         </dependency>
 
         <!-- test -->
